+++ src/orpg/server/net/handlers/EditorRequestSegmentHandler.java
@@ -9,8 +9,8 @@ public class EditorRequestSegmentHandler implements ServerPacketHandler {
 	@Override
 	public void handle(ServerReceivedPacket packet, BaseServer baseServer) {
 		int id = packet.getByteBuffer().getInt();
-		int x = packet.getByteBuffer().getInt();
-		int y = packet.getByteBuffer().getInt();
+		short x = packet.getByteBuffer().getShort();
+		short y = packet.getByteBuffer().getShort();
 
 		baseServer.sendPacket(new EditorMapSegmentDataPacket(packet
 				.getSession(), id, baseServer.getMapController().getSegment(id, x,
